requirements.txt
asgiref==3.8.1
astroid==3.3.5
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==3.3.2
cmake==3.30.4
colorama==0.4.6
coverage==7.6.1
cryptography==36.0.0
defusedxml==0.8.0rc2
dependency-injector==4.37.0
dill==0.3.9
distlib==0.3.8
Django==5.1.2
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
exceptiongroup==1.2.2
filelock==3.16.1
h11==0.12.0
idna==3.3
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mysqlclient==2.2.4
oauthlib==3.2.2
outcome==1.1.0
packaging==24.1
pipenv==2024.1.0
platformdirs==4.3.6
pluggy==1.5.0
py==1.10.0
pycparser==2.21
PyJWT==2.9.0
pylint==3.3.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pytest==8.3.3
pytest-django==4.9.0
pytest-random==0.2
pytest-random-order==1.0.4
python3-openid==3.2.0
pytz==2022.6
requests==2.32.3
requests-oauthlib==2.0.0
selenium==4.1.0
singleton-decorator==1.0.0
six==1.16.0
sniffio==1.2.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sortedcontainers==2.4.0
sqlparse==0.5.1
supervisor==4.1.0
toml==0.10.2
tomli==2.0.2
tomlkit==0.13.2
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.7
virtualenv==20.26.6
wsproto==1.0.0
